backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.9
python-dotenv==1.1.1
groq==0.31.1
PyMuPDF==1.26.4
python-docx==1.2.0